+++ server/src/main/java/com/example/hackathon/account/infrastructure/MockSHAccountReaderAdapter.java
@@ -3,7 +3,7 @@ package com.example.hackathon.account.infrastructure;
 import com.example.hackathon.account.domain.*;
 import com.example.hackathon.account.domain.shResponse.SHAccountREC;
 import com.example.hackathon.account.domain.shResponse.SHTransactionHistoryREC;
-import com.example.hackathon.common.shResponse.SHApiRECResponse;
+import com.example.hackathon.common.shDto.SHApiRECResponse;
 import com.example.hackathon.common.mock.MockApiLoader;
 import com.fasterxml.jackson.core.type.TypeReference;
 import lombok.RequiredArgsConstructor;
